Move EditAlg into edit package
